pyproject.toml
[project]
name = "solugen-assignment"
version = "0.1.0"
description = "RAG retrieval component (no LLM generation) for Solugen assignment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
  "chromadb>=0.5.0",
  "openai>=1.0.0",
  "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/solugen_assignment"]

[tool.ruff]
line-length = 100
target-version = "py310"


